backend/services/web_search_service/requirements.txt
langchain-groq
langchain-core
fastapi
uvicorn
tavily-python
dotenv
